# A few performance improvements