--- a/master_changes.md
+++ b/master_changes.md
@@ -161,3 +161,4 @@ users)
 ## opam-core
   * `OpamCmdliner` was added. It is accessible through a new `opam-core.cmdliner` sub-library [#6755 @kit-ty-kate]
   * `OpamUrl`: rename and expose `local_path` as `looks_like_local_path` [#5979 @kit-ty-kate]
+  * `OpamCompat.Gc.ramp_up`: was added [#6515 @dra27]