--- a/master_changes.md
+++ b/master_changes.md
@@ -115,6 +115,7 @@ users)
 ## Shell
 
 ## Internal
+  * Improve cache-loading performance when using OCaml >= 5.4 by using `Gc.ramp_up` [#6515 @dra27]
 
 ## Internal: Unix