--- a/src/core/opamCompat.ml
+++ b/src/core/opamCompat.ml
@@ -261,3 +261,12 @@ module Pair = struct
   let equal eq1 eq2 (x1, y1) (x2, y2) =
     eq1 x1 x2 && eq2 y1 y2
 end
+
+module Int = struct
+  [@@@warning "-32"]
+
+  (* NOTE: OCaml >= 4.13 *)
+  let min : int -> int -> int = Stdlib.min
+
+  include Stdlib.Int
+end