--- a/src/core/opamCompat.mli
+++ b/src/core/opamCompat.mli
@@ -8,6 +8,12 @@
 (*                                                                        *)
 (**************************************************************************)
 
+module Gc : sig
+  (* NOTE: OCaml >= 5.4; we don't expose the suspended_collection_work as
+           it's not needed by opam and it complicates the backport *)
+  val ramp_up : (unit -> 'a) -> 'a
+end
+
 module String : sig
   (* NOTE: OCaml >= 4.13 *)
   val exists: (char -> bool) -> string -> bool