--- a/src/core/opamStd.ml
+++ b/src/core/opamStd.ml
@@ -632,25 +632,25 @@ module OpamString = struct
         f acc next current last normal in
     f [] 0 "" 0 true
 
+  let rec compare_case_aux len s1 s2 i =
+    if i < (len : int) then
+      let c1 = String.unsafe_get s1 i and c2 = String.unsafe_get s2 i in
+      match Char.compare (Char.lowercase_ascii c1) (Char.lowercase_ascii c2)
+      with
+      | 0 ->
+        (match Char.compare c1 c2 with
+         | 0 -> compare_case_aux len s1 s2 (i+1)
+         | c -> c)
+      | c -> c
+    else
+      let l1 = String.length s1 and l2 = String.length s2 in
+      if l1 < (l2 : int) then -1
+      else if l1 > (l2 : int) then 1
+      else 0
+
   let compare_case s1 s2 =
-    let l1 = String.length s1 and l2 = String.length s2 in
-    let len = min l1 l2 in
-    let rec aux i =
-      if i < len then
-        let c1 = s1.[i] and c2 = s2.[i] in
-        match Char.compare (Char.lowercase_ascii c1) (Char.lowercase_ascii c2)
-        with
-        | 0 ->
-          (match Char.compare c1 c2 with
-           | 0 -> aux (i+1)
-           | c -> c)
-        | c -> c
-      else
-        if l1 < l2 then -1
-        else if l1 > l2 then 1
-        else 0
-    in
-    aux 0
+    let len = OpamCompat.Int.min (String.length s1) (String.length s2) in
+    compare_case_aux len s1 s2 0
 
   let is_prefix_of ~from ~full s =
     let length_s = String.length s in